Move encrypt to OpenPGPKey
